stegashield_model - Copy/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
python-multipart>=0.0.9
numpy>=1.24
Pillow>=10.0
opencv-python>=4.8
pywavelets>=1.5

